+++ examples/untyped-react-example/src/VariantsWithPayload.gen.js
@@ -12,15 +12,49 @@ const $$toJS346759412 = {"0": "A", "1": "B", "2": "C"};
 
 const $$toRE346759412 = {"A": 0, "B": 1, "C": 2};
 
+const $$toJS435467058 = {"a": "a", "b": "bRenamed", "True": true, "Twenty": 20, "Half": 0.5};
+
+const $$toRE435467058 = {"a": "a", "bRenamed": "b", "true": "True", "20": "Twenty", "0.5": "Half"};
+
 import * as VariantsWithPayloadBS from './VariantsWithPayload.bs';
 
-export const testWithPayload = VariantsWithPayloadBS.testWithPayload;
+export const testWithPayload = function (Arg1) {
+  const result = VariantsWithPayloadBS.testWithPayload(typeof(Arg1) === 'object'
+    ? {NAME: "c", VAL: Arg1.VAL}
+    : $$toRE435467058[Arg1.toString()]);
+  return typeof(result) === 'object'
+    ? {NAME:"c", VAL:result.VAL}
+    : $$toJS435467058[result]
+};
 
-export const printVariantWithPayload = VariantsWithPayloadBS.printVariantWithPayload;
+export const printVariantWithPayload = function (Arg1) {
+  const result = VariantsWithPayloadBS.printVariantWithPayload(typeof(Arg1) === 'object'
+    ? {NAME: "c", VAL: Arg1.VAL}
+    : $$toRE435467058[Arg1.toString()]);
+  return result
+};
 
-export const testManyPayloads = VariantsWithPayloadBS.testManyPayloads;
+export const testManyPayloads = function (Arg1) {
+  const result = VariantsWithPayloadBS.testManyPayloads(Arg1.NAME==="oneRenamed"
+    ? {NAME: "one", VAL: Arg1.VAL}
+    : Arg1.NAME===2
+    ? {NAME: "two", VAL: Arg1.VAL}
+    : {NAME: "three", VAL: Arg1.VAL});
+  return result.NAME==="one"
+    ? {NAME:"oneRenamed", VAL:result.VAL}
+    : result.NAME==="two"
+    ? {NAME:2, VAL:result.VAL}
+    : {NAME:"three", VAL:result.VAL}
+};
 
-export const printManyPayloads = VariantsWithPayloadBS.printManyPayloads;
+export const printManyPayloads = function (Arg1) {
+  const result = VariantsWithPayloadBS.printManyPayloads(Arg1.NAME==="oneRenamed"
+    ? {NAME: "one", VAL: Arg1.VAL}
+    : Arg1.NAME===2
+    ? {NAME: "two", VAL: Arg1.VAL}
+    : {NAME: "three", VAL: Arg1.VAL});
+  return result
+};
 
 export const testSimpleVariant = function (Arg1) {
   const result = VariantsWithPayloadBS.testSimpleVariant($$toRE346759412[Arg1]);